Maintained Project Structure - routes, controllers,utils
