+++ user-auth-aws-cognito/src/utils/expressUtils.js
@@ -1,0 +1,4 @@
+const express = require('express');
+module.exports={
+    express
+}
