+++ user-auth-aws-cognito/tsconfig.app.json
@@ -3,8 +3,8 @@
   "compilerOptions": {
     "outDir": "./dist/out-tsc",
     "module": "commonjs",
-    "types": ["node"]
+    "types": ["node","express"]
   },
   "exclude": ["jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts"],
-  "include": ["src/**/*.ts", "src/main.js","src/utils/validate.js"]
+  "include": ["src/**/*.ts", "src/main.js","src/**/**.js"]
 }
