--- user-auth-aws-cognito/utils/validate.js
@@ -1,24 +1,0 @@
-const sendResponse = (statusCode, body) => {
-    const response = {
-        statusCode: statusCode,
-        body: JSON.stringify(body),
-        headers: {
-            'Content-Type': 'application/json',
-            'Access-Control-Allow-Origin': '*',
-            'Access-Control-Allow-Credentials': true
-        }
-    }
-    return response
-}
-
-const validateInput = (data) => {
-   // const body = JSON.parse(data);
-    const { email, password } = data
-    if (!email || !password || password.length < 6)
-        return false
-    return true
-}
-
-module.exports = {
-    sendResponse, validateInput
-};
